backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# PostgreSQL
asyncpg==0.29.0

# MongoDB
motor==3.4.0

# Redis
redis==5.0.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP / AI
httpx==0.27.0
openai==1.25.0

# Scheduler
apscheduler==3.10.4

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
